# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Reserva de Laboratórios e Salas",
    "version": "1.0.0",
    "description": "Importe no Insomnia. Cadastre usuário, confira login e configure Basic Auth. Use IDs retornados nas consultas. Datas DD/MM/AAAA e horários HH:MM:SS. Testes ainda pendentes."
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "security": [
    {
      "basicAuth": []
    }
  ],
  "paths": {
...
    "/api/status-reservas": {
      "get": {
        "summary": "Consultar status-reservas",
        "tags": [
          "status-reservas"
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StatusView"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Autenticação inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Registro não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflito",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Cadastrar status-reservas",
        "tags": [
          "status-reservas"
        ],
        "responses": {
          "201": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusView"
                }
              }
            },
            "headers": {
              "Location": {
                "description": "URI do registro criado",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Autenticação inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Registro não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflito",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusInput"
              }
            }
          }
        }
      }
    },
    "/api/status-reservas/{codigo}": {
      "get": {
        "summary": "Buscar por código",
        "tags": [
          "status-reservas"
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusView"
                }
              }
            }
          },
          "401": {
            "description": "Autenticação inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Registro não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflito",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "codigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          }
        ]
      },
      "put": {
        "summary": "Atualizar cadastro completo",
        "tags": [
          "status-reservas"
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusView"
                }
              }
            }
          },
          "401": {
            "description": "Autenticação inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Registro não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflito",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusInput"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "codigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          }
        ]
      },
      "delete": {
        "summary": "Excluir registro (reserva ativa respeita prazo de cancelamento)",
        "tags": [
          "status-reservas"
        ],
        "responses": {
          "204": {
            "description": "Sucesso"
          },
          "401": {
            "description": "Autenticação inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Registro não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflito",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "codigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          }
        ]
      }
    },
...
    "/api/reservas/{codigo}": {
      "get": {
        "summary": "Buscar por código",
        "tags": [
          "reservas"
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservaView"
                }
              }
            }
          },
          "401": {
            "description": "Autenticação inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Registro não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflito",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "codigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          }
        ]
      },
      "put": {
        "summary": "Atualizar cadastro completo",
        "tags": [
          "reservas"
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservaView"
                }
              }
            }
          },
          "401": {
            "description": "Autenticação inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Registro não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflito",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReservaInput"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "codigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          }
        ]
      },
      "delete": {
        "summary": "Excluir registro (reserva ativa respeita prazo de cancelamento)",
        "tags": [
          "reservas"
        ],
        "responses": {
          "204": {
            "description": "Sucesso"
          },
          "401": {
            "description": "Autenticação inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Registro não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflito",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "codigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    },
    "schemas": {
...
      "UsuarioView": {
        "type": "object",
        "properties": {
          "codigo": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "cpf": {
            "type": "string",
            "minLength": 11,
            "maxLength": 11,
            "example": "52998224725"
          },
          "nomeCompleto": {
            "type": "string",
            "minLength": 10,
            "maxLength": 80,
            "example": "Professor de Teste"
          },
          "dataAniversario": {
            "type": "string",
            "pattern": "^\\d{2}/\\d{2}/\\d{4}$",
            "example": "15/03/1990",
            "description": "DD/MM/AAAA; use uma data futura na reserva"
          },
          "celular": {
            "type": "string",
            "maxLength": 20,
            "example": "11999999999"
          },
          "email": {
            "type": "string",
            "format": "email",
            "minLength": 15,
            "maxLength": 80,
            "example": "professor.teste@example.com"
          },
          "login": {
            "type": "string",
            "pattern": "^[A-Za-z0-9._-]{3,80}$",
            "example": "professor.teste"
          }
        },
        "required": [
          "codigo",
          "cpf",
          "nomeCompleto",
          "dataAniversario",
          "celular",
          "email",
          "login"
        ],
        "additionalProperties": false
      },
...
      "StatusInput": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "minLength": 3,
            "maxLength": 20,
            "example": "MANUTENCAO"
          }
        },
        "required": [
          "nome"
        ],
        "additionalProperties": false
      },
      "StatusView": {
        "type": "object",
        "properties": {
          "codigo": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "nome": {
            "type": "string"
          }
        },
        "required": [
          "codigo",
          "nome"
        ],
        "additionalProperties": false
      },
...
      "RecursoView": {
        "type": "object",
        "properties": {
          "codigo": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "tipo": {
            "type": "string",
            "enum": [
              "LABORATORIO",
              "SALA"
            ]
          },
          "nome": {
            "type": "string"
          },
          "capacidade": {
            "type": "integer"
          },
          "localizacao": {
            "type": "string"
          },
          "statusRecurso": {
            "$ref": "#/components/schemas/StatusView"
          }
        },
        "required": [
          "codigo",
          "tipo",
          "nome",
          "capacidade",
          "localizacao",
          "statusRecurso"
        ],
        "additionalProperties": false
      },
      "ReservaInput": {
        "type": "object",
        "properties": {
          "dataInicial": {
            "type": "string",
            "pattern": "^\\d{2}/\\d{2}/\\d{4}$",
            "example": "30/09/2026",
            "description": "DD/MM/AAAA; use uma data futura na reserva"
          },
          "dataFinal": {
            "type": "string",
            "pattern": "^\\d{2}/\\d{2}/\\d{4}$",
            "example": "30/09/2026",
            "description": "DD/MM/AAAA; use uma data futura na reserva"
          },
          "horaInicial": {
            "type": "string",
            "pattern": "^\\d{2}:\\d{2}:\\d{2}$",
            "example": "10:00:00"
          },
          "horaFinal": {
            "type": "string",
            "pattern": "^\\d{2}:\\d{2}:\\d{2}$",
            "example": "11:00:00"
          },
          "usuarioCodigo": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "recursoCodigo": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "statusReservaCodigo": {
            "type": "integer",
            "format": "int64",
            "example": 1
          }
        },
        "required": [
          "dataInicial",
          "dataFinal",
          "horaInicial",
          "horaFinal",
          "usuarioCodigo",
          "recursoCodigo",
          "statusReservaCodigo"
        ],
        "additionalProperties": false
      },
      "ReservaView": {
        "type": "object",
        "properties": {
          "codigo": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "dataInicial": {
            "type": "string",
            "pattern": "^\\d{2}/\\d{2}/\\d{4}$",
            "example": "30/09/2026",
            "description": "DD/MM/AAAA; use uma data futura na reserva"
          },
          "dataFinal": {
            "type": "string",
            "pattern": "^\\d{2}/\\d{2}/\\d{4}$",
            "example": "30/09/2026",
            "description": "DD/MM/AAAA; use uma data futura na reserva"
          },
          "horaInicial": {
            "type": "string",
            "pattern": "^\\d{2}:\\d{2}:\\d{2}$",
            "example": "10:00:00"
          },
          "horaFinal": {
            "type": "string",
            "pattern": "^\\d{2}:\\d{2}:\\d{2}$",
            "example": "11:00:00"
          },
          "usuario": {
            "$ref": "#/components/schemas/UsuarioView"
          },
          "recurso": {
            "$ref": "#/components/schemas/RecursoView"
          },
          "statusReserva": {
            "$ref": "#/components/schemas/StatusView"
          }
        },
        "required": [
          "codigo",
          "dataInicial",
          "dataFinal",
          "horaInicial",
          "horaFinal",
          "usuario",
          "recurso",
          "statusReserva"
        ],
        "additionalProperties": false
      },
      "ApiError": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "integer"
          },
          "mensagem": {
            "type": "string"
          },
          "caminho": {
            "type": "string"
          },
          "erros": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "campo": {
                  "type": "string"
                },
                "mensagem": {
                  "type": "string"
                }
              },
              "required": [
                "campo",
                "mensagem"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "timestamp",
          "status",
          "mensagem",
          "caminho",
          "erros"
        ],
        "additionalProperties": false
      }
    }
  }
}